fix(event): use correct types for window proc params

+++ src/event/Window.hpp
@@ -3,6 +3,6 @@
 
 #include <cstdint>
 
-int32_t OsWindowProc(void* window, uint32_t message, uint32_t wparam, int32_t lparam);
+int32_t OsWindowProc(void* window, uint32_t message, uintptr_t wparam, intptr_t lparam);
 
 #endif
